[Silver II] Title: 나무 자르기, Time: 460 ms, Memory: 29724 KB -BaekjoonHub

+++ Java/백준/Silver/14247. 나무 자르기/나무 자르기.java
@@ -15,12 +15,12 @@ class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int N = Integer.parseInt(st.nextToken());
 
-        long[] H = new long[N];
+        int[] H = new int[N];
         IV[] P = new IV[N];
 
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; ++i) {
-            long val = Long.parseLong(st.nextToken());
+            int val = Integer.parseInt(st.nextToken());
             H[i] = val;
         }
